integrate firebase for rooms and objects
