+++ src/components/room/Environment.js
@@ -1,6 +1,11 @@
 import * as THREE from "three";
 
-export const addEnvironment = (scene) => {
+export const addEnvironment = ({
+  scene,
+  toDispose,
+  toRemove,
+  raycastHelperObjects,
+}) => {
   //room shape
   const wallWidth = 10;
   const wallHeight = 8;
@@ -10,6 +15,7 @@ export const addEnvironment = (scene) => {
     wallWidth,
     wallHeight
   );
+  toDispose.push(wallGeo);
 
   const backWallWidth = 20;
   const backWallHeight = 8;
@@ -19,6 +25,7 @@ export const addEnvironment = (scene) => {
     backWallWidth,
     backWallHeight
   );
+  toDispose.push(backWallGeo);
 
   const floorGeo = new THREE.PlaneBufferGeometry(
     backWallWidth,
@@ -26,21 +33,25 @@ export const addEnvironment = (scene) => {
     backWallWidth,
     wallWidth
   );
+  toDispose.push(backWallGeo);
 
-  var wallMaterial = new THREE.MeshBasicMaterial({
+  const wallMaterial = new THREE.MeshBasicMaterial({
     color: 0xffffff,
     opacity: 0.9,
     side: THREE.DoubleSide,
     wireframe: true,
     transparent: true,
   });
-  var floorMaterial = new THREE.MeshBasicMaterial({
+  toDispose.push(wallMaterial);
+
+  const floorMaterial = new THREE.MeshBasicMaterial({
     color: 0x32a852,
     opacity: 0.9,
     side: THREE.DoubleSide,
     wireframe: true,
     transparent: true,
   });
+  toDispose.push(floorMaterial);
 
   let gap = 0.5;
 
@@ -49,33 +60,31 @@ export const addEnvironment = (scene) => {
   leftWall.position.x = backWallWidth / 2 + gap;
   leftWall.position.z = wallWidth / 2 + gap;
   scene.add(leftWall);
+  toRemove.push(leftWall);
 
   let rightWall = new THREE.Mesh(wallGeo, wallMaterial);
   rightWall.rotation.y = Math.PI / 2;
   rightWall.position.x = -backWallWidth / 2 - gap;
   rightWall.position.z = wallWidth / 2 + gap;
   scene.add(rightWall);
+  toRemove.push(rightWall);
 
   let backWall = new THREE.Mesh(backWallGeo, wallMaterial);
   scene.add(backWall);
+  toRemove.push(backWall);
 
   let floor = new THREE.Mesh(floorGeo, floorMaterial);
   floor.rotation.x = -Math.PI / 2;
+  toRemove.push(floor);
 
   floor.position.y = -wallHeight / 2 - gap;
   floor.position.z = wallWidth / 2 + gap;
 
   scene.add(floor);
 
+  raycastHelperObjects.push(floor, leftWall, backWall, rightWall);
+
   return {
     floor,
-    backWall,
-    rightWall,
-    leftWall,
-    wallGeo,
-    floorGeo,
-    backWallGeo,
-    wallMaterial,
-    floorMaterial,
   };
 };
